proxylib: move API implementation into libcilium subpackage
The goal is to make `CGO_ENABLED=0 go test ./...` work on the whole repository.
This commit moves the libcilium implementation and tests to a subpackage that
can be made free of 'import "C"' statements, reducing `./proxylib` to be a cgo
export shim only, without needing any (Cgo-enabled) tests of its own.

This also moves the tests out of package main, which currently annoyingly
requires running `go test ./proxylib/*.go` for invoking the tests.
